requirements.txt
blessed==1.20.0
boto3==1.34.93
botocore==1.34.162
certifi==2024.8.30
cffi==1.17.1
chalice==1.31.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
gcloud==0.18.3
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
inquirer==2.10.1
jmespath==1.0.1
jwcrypto==1.5.6
oauth2client==4.1.3
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.0
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-jwt==4.1.0
PyYAML==6.0.2
readchar==4.2.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.10.4
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
